+++ TranslationsHelper/Package/CHANGELOG.md
@@ -1,1 +1,13 @@
+* 1.0.5 ->
+    * fixed issues not providing translations
+    * added types Character, Beehive, CookingStation, Fermenter, Incinerator, MapTable, OfferingBowl, SapCollector,
+      Smelter
+    * string outputs in translation files will be sanity checked for non-word special characters (except "_")
+    * if tokens were not changed in custom items from vanilla tokens, will also print the vanilla tokens
+* 1.0.4 ->
+    * updated Jotunn dependency
+    * updated for Valheim version 0.217.14
+* 1.0.3 -> updated Jotunn dependency
+* 1.0.2 -> changed from writing .yaml to .yml (if that matters)
+* 1.0.1 -> fixed duplicate mod name key error
 * 1.0.0 -> first release